+++ tests/Patch/Manifest/ManifestCheckerTest.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace Meteor\Patch\Manifest;
+
+use Meteor\IO\NullIO;
+use org\bovigo\vfs\vfsStream;
+
+class ManifestCheckerTest extends \PHPUnit_Framework_TestCase
+{
+    private $manifestChecker;
+
+    public function setUp()
+    {
+        $this->manifestChecker = new ManifestChecker(new NullIO());
+
+        vfsStream::setup('root', null, [
+            'a.txt' => 'a',
+            'b.txt' => 'b',
+            'c.txt' => 'c',
+        ]);
+    }
+
+    public function testReturnsFalseIfManifestFileDoesNotExist()
+    {
+        $this->assertFalse($this->manifestChecker->check(vfsStream::url('root')));
+    }
+
+    public function testReturnsFalseIfManifestFileCouldNotBeParsed()
+    {
+        file_put_contents(vfsStream::url('root/' . ManifestChecker::MANIFEST_FILENAME), '!!');
+
+        $this->assertFalse($this->manifestChecker->check(vfsStream::url('root')));
+    }
+
+    public function testReturnsFalseIfThePatchCannotBeVerifiedDueToMissingFile()
+    {
+        $manifest = [
+            'd.txt' => 'hash',
+        ];
+
+        file_put_contents(vfsStream::url('root/' . ManifestChecker::MANIFEST_FILENAME), json_encode($manifest));
+
+        $this->assertFalse($this->manifestChecker->check(vfsStream::url('root')));
+    }
+
+    public function testReturnsFalseIfThePatchCannotBeVerifiedDueToInvalidFileHash()
+    {
+        $manifest = [
+            'c.txt' => 'hash',
+        ];
+
+        file_put_contents(vfsStream::url('root/' . ManifestChecker::MANIFEST_FILENAME), json_encode($manifest));
+
+        $this->assertFalse($this->manifestChecker->check(vfsStream::url('root')));
+    }
+
+    public function testReturnsTrueIfThePatchIsVerified()
+    {
+        $manifest = [
+            'a.txt' => '86f7e437faa5a7fce15d1ddcb9eaeaea377667b8',
+            'b.txt' => 'e9d71f5ee7c92d6dc9e92ffdad17b8bd49418f98',
+            'c.txt' => '84a516841ba77a5b4648de2cd0dfcb30ea46dbb4',
+        ];
+
+        file_put_contents(vfsStream::url('root/' . ManifestChecker::MANIFEST_FILENAME), json_encode($manifest));
+
+        $this->assertTrue($this->manifestChecker->check(vfsStream::url('root')));
+    }
+}